pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shaytan-machine"
version = "0.1.0"
description = "MAGAMAX Shaytan Machine modular monolith"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "boto3>=1.35.0",
  "dramatiq>=1.17.1",
  "duckdb>=1.1.3",
  "fastapi>=0.115.4",
  "httpx>=0.28.1",
  "opentelemetry-api>=1.30.0",
  "opentelemetry-exporter-otlp>=1.30.0",
  "opentelemetry-instrumentation-fastapi>=0.51b0",
  "opentelemetry-instrumentation-sqlalchemy>=0.51b0",
  "opentelemetry-sdk>=1.30.0",
  "openpyxl>=3.1.5",
  "pandas>=2.2.3",
  "psycopg[binary]>=3.2.3",
  "pyarrow>=18.0.0",
  "pydantic-settings>=2.6.1",
  "pyjwt>=2.9.0",
  "python-multipart>=0.0.12",
  "redis>=5.2.0",
  "sentry-sdk[fastapi]>=2.18.0",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0"
]

[project.optional-dependencies]
dev = [
  "black>=24.10.0",
  "mypy>=1.13.0",
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.7.4"
]

[tool.setuptools]
packages = []

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["B008", "B009", "E501", "UP037", "UP046"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
disallow_untyped_defs = true
ignore_missing_imports = true
mypy_path = ["apps/api", "apps/worker"]

[tool.pytest.ini_options]
pythonpath = ["apps/api", "apps/worker", "."]
testpaths = ["tests"]
